\name{data}
\alias{data}
\alias{BanInf}
\alias{dat}

\docType{data}
\title{empirical datasets}
\description{

BanInf: Productivity of crosses involving Tribolium castaneum[1]

}
\usage{data(BanInf)}

\source{
[1] Demuth, J.P., Evolution of Hybrid Incompatibility in the beetle Tribolium Castaneum, in Biology. 2004, Indiana University: Bloomington. p. 152.

[2] Lynch, M., and B. Walsh, 1998 Lynch, M., & Walsh, B. (1998). Genetics and analysis of quantitative traits. Sinauer Associates, Inc., Sunderland, Massachusetts.

[3] Miller, G. T., Starmer, W. T., & S. Pitnick 2003. Quantitative genetic analysis of among-population variation in sperm and female sperm-storage organ length in Drosophila mojavensis. Genetical research, 81(03), 213-220.

}
\examples{
data(ban.osa)
data(dar.bho)
data(per.inf)
data(sin.cro)
data(PH)
data(SL)
data(SR)
}
\keyword{datasets}