\name{EvaluateModel}
\alias{EvaluateModel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plots the results of a single model}
\description{
Plot the parameter estimates derived from a single selected model.}
\usage{EvaluateModel(data, model, cex.axis=1, cex.names=1, cex.main=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{This is the observed data used in the function AnalyzeCrossesMM}
  \item{model}{This is a "genarch" object returned by the function AnalyzeCrossesMM}
  \item{cex.axis}{expansion factor for Y axis}
  \item{cex.names}{expansion factor for CGE names on x axis}
  \item{cex.main}{expansion factor for main title}
}
\author{
Heath Blackmon and Jeffery P. Demuth}
