\name{SAGA-package}
\alias{SAGA-package}
\alias{SAGA}
\docType{package}
\title{
Software for the Analysis of Genetic Architecture

SAGA}
\description{
Provides multimodel parameter estimates of genetic architecture incorporating the impact of environmental and sex effects.  Uses weighted least squares to solve all possible models.  Chooses the model or models that are most appropriate based on AICc weights and produces model weighted parameter estimates and unconditional standard errors.  A function to visualize the distribution of model fits across model space allows users to graphically see how well defined the "true" model is with their dataset.
}
\details{
\tabular{ll}{
Package: \tab SAGA\cr
Version: \tab 2.0.0\cr
}
}
\author{
Heath Blackmon\cr 
Jeffery P. Demuth\cr
Andrew Armstrong\cr
Jonathon Leach\cr
Correspondence: Heath Blackmon at coleoguy@gmail.com}
\references{
Blackmon, H. and J. Demuth. 2016. An information-theoretic approach to estimating the composite genetic effects contributing to variation among generation means: Moving beyond the joint-scaling test for line cross analysis. Evolution 70:2 420-432.\cr\cr
Lynch, M. and B. Walsh. 1998. Genetics and analysis of quantitative traits. Sinauer Associates, Inc., Sunderland, Massachusetts.\cr\cr
Burnham, K. P. and D. R. Anderson. 2002. Model selection and multimodel inference: a practical information-theoretic approach. Springer Science & Business Media.}
