% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGA2-package.R
\docType{package}
\name{SAGA2-package}
\alias{SAGA2}
\alias{SAGA2-package}
\title{SAGA2: Software for the Analysis of Genetic Architecture}
\description{
Implements an information-theoretic approach to the analysis of line cross data providing model averaged results of parameter estimates and unconditional standard errors. Uses AICc-based model selection to explore all possible models of composite genetic effects contributing to variation among generation means. Also includes functions to visualize model space, produce custom plots of multi-model inference results, and traditional line cross analysis plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coleoguy/SAGA2}
  \item Report bugs at \url{https://github.com/coleoguy/SAGA2/issues}
}

}
\author{
\strong{Maintainer}: Heath Blackmon \email{coleoguy@gmail.com} (\href{https://orcid.org/0000-0002-5433-4036}{ORCID})

Authors:
\itemize{
  \item Jeffery P. Demuth
}

Other contributors:
\itemize{
  \item Andrew Armstrong [contributor]
  \item Nathan Anderson [contributor]
  \item Jonathon Leach [contributor]
}

}
\keyword{internal}
